Seat 관련된 Service 생성
